+++ tests/decay/test_decay.py
@@ -5,6 +5,7 @@
 # or https://github.com/scikit-hep/decaylanguage for details.
 
 
+import pytest
 from pytest import approx
 
 from decaylanguage.decay.decay import DaughtersDict
@@ -71,16 +72,6 @@ def test_DecayMode_constructor_simple():
     assert dm.metadata == dict(model="", model_params="")
 
 
-def test_DecayMode_constructor_from_pdgids():
-    dm = DecayMode.from_pdgids(
-        0.5,
-        [321, -321],
-        model="TAUHADNU",
-        model_params=[-0.108, 0.775, 0.149, 1.364, 0.400],
-    )
-    assert dm.daughters == DaughtersDict("K+ K-")
-
-
 def test_DecayMode_constructor_with_model_info():
     dd = DaughtersDict("pi- pi0 nu_tau")
     dm = DecayMode(
@@ -92,7 +83,7 @@ def test_DecayMode_constructor_with_model_info():
     }
 
 
-def test_DecayMode_constructor_with_user_model_info():
+def test_DecayMode_constructor_with_user_metadata():
     dd = DaughtersDict("K+ K-")
     dm = DecayMode(0.5, dd, model="PHSP", study="toy", year=2019)
     assert dm.metadata == {
@@ -101,6 +92,23 @@ def test_DecayMode_constructor_with_user_model_info():
         "study": "toy",
         "year": 2019,
     }
+
+
+def test_DecayMode_constructor_from_pdgids_default():
+    dm = DecayMode.from_pdgids()
+    assert dm.bf == 0
+    assert dm.daughters == DaughtersDict()
+    assert dm.metadata == dict(model="", model_params="")
+
+
+def test_DecayMode_constructor_from_pdgids():
+    dm = DecayMode.from_pdgids(
+        0.5,
+        [321, -321],
+        model="TAUHADNU",
+        model_params=[-0.108, 0.775, 0.149, 1.364, 0.400],
+    )
+    assert dm.daughters == DaughtersDict("K+ K-")
 
 
 def test_DecayMode_constructor_from_dict():
@@ -119,7 +127,7 @@ def test_DecayMode_describe_simple():
     assert "Decay model: TAUHADNU [-0.108, 0.775, 0.149, 1.364, 0.4]" in dm.describe()
 
 
-def test_DecayMode_describe_with_extra_info():
+def test_DecayMode_describe_with_user_metadata():
     dd = DaughtersDict("K+ K-")
     dm = DecayMode(1.0e-6, dd, model="PHSP", study="toy", year=2019)
     assert "Extra info:" in dm.describe()
@@ -153,20 +161,24 @@ def test_DecayMode_number_of_final_states():
     assert len(dm) == 4
 
 
-dm1 = DecayMode(0.0124, "K_S0 pi0", model="PHSP")
-dm2 = DecayMode(0.692, "pi+ pi-")
-dm3 = DecayMode(0.98823, "gamma gamma")
-dc = DecayChain("D0", {"D0": dm1, "K_S0": dm2, "pi0": dm3})
+@pytest.fixture()
+def dc():
+    dm1 = DecayMode(0.0124, "K_S0 pi0", model="PHSP")
+    dm2 = DecayMode(0.692, "pi+ pi-")
+    dm3 = DecayMode(0.98823, "gamma gamma")
+    return DecayChain("D0", {"D0": dm1, "K_S0": dm2, "pi0": dm3})
 
 
-dm1 = DecayMode(0.6770, "D0 pi+")
-dm2 = DecayMode(0.0124, "K_S0 pi0")
-dm3 = DecayMode(0.692, "pi+ pi-")
-dm4 = DecayMode(0.98823, "gamma gamma")
-dc2 = DecayChain("D*+", {"D*+": dm1, "D0": dm2, "K_S0": dm3, "pi0": dm4})
+@pytest.fixture()
+def dc2():
+    dm1 = DecayMode(0.6770, "D0 pi+")
+    dm2 = DecayMode(0.0124, "K_S0 pi0")
+    dm3 = DecayMode(0.692, "pi+ pi-")
+    dm4 = DecayMode(0.98823, "gamma gamma")
+    return DecayChain("D*+", {"D*+": dm1, "D0": dm2, "K_S0": dm3, "pi0": dm4})
 
 
-def test_DecayChain_constructor_subdecays():
+def test_DecayChain_constructor_subdecays(dc):
     assert len(dc.decays) == 3
     assert dc.mother == "D0"
 
@@ -206,7 +218,7 @@ def test_DecayChain_constructor_from_dict():
     assert DecayChain.from_dict(dc_dict).to_dict() == dc_dict
 
 
-def test_DecayChain_to_dict():
+def test_DecayChain_to_dict(dc2):
     assert dc2.to_dict() == {
         "D*+": [
             {
@@ -252,16 +264,47 @@ def test_DecayChain_to_dict():
     }
 
 
-def test_DecayChain_properties():
+def test_DecayChain_properties(dc):
     assert dc.bf == 0.0124
     assert dc.visible_bf == approx(0.008479803984)
 
 
-def test_DecayChain_flatten():
-    dc_flatten = dc2.flatten()
-    assert dc_flatten.mother == dc2.mother
-    assert dc_flatten.bf == dc_flatten.visible_bf
+def test_DecayChain_flatten(dc2):
+    dc2_flatten = dc2.flatten()
+    assert dc2_flatten.mother == dc2.mother
+    assert dc2_flatten.bf == dc2_flatten.visible_bf
+    assert dc2_flatten.decays[dc2.mother].daughters == DaughtersDict(
+        ["gamma", "gamma", "pi+", "pi+", "pi-"]
+    )
 
 
-def test_DecayChain_string_repr():
+def test_DecayChain_flatten_complex():
+    dm1 = DecayMode(1.0, "D0 K_S0 pi+ pi0")
+    dm2 = DecayMode(0.0124, "K_S0 pi0 pi0")
+    dm3 = DecayMode(0.692, "pi+ pi-")
+    dm4 = DecayMode(0.98823, "gamma gamma")
+    dc = DecayChain("X+", {"X+": dm1, "D0": dm2, "K_S0": dm3, "pi0": dm4})
+    dc_flatten = dc.flatten()
+    assert dc_flatten.decays[dc_flatten.mother].daughters == DaughtersDict(
+        {"gamma": 6, "pi+": 3, "pi-": 2}
+    )
+    assert dc_flatten.bf == approx(1.0 * 0.0124 * (0.692 ** 2) * (0.98823 ** 3))
+
+
+def test_DecayChain_flatten_with_stable_particles():
+    dm1 = DecayMode(0.5, "D0 anti-D0 pi+ pi0 pi0")
+    dm2 = DecayMode(0.0124, "K_S0 pi0")
+    dm3 = DecayMode(0.692, "pi+ pi-")
+    dm4 = DecayMode(0.98823, "gamma gamma")
+    dc = DecayChain(
+        "X+", {"X+": dm1, "D0": dm2, "anti-D0": dm2, "K_S0": dm3, "pi0": dm4}
+    )
+    dc_flatten = dc.flatten(stable_particles=["pi0"])
+    assert dc_flatten.decays[dc_flatten.mother].daughters == DaughtersDict(
+        {"pi0": 4, "pi+": 3, "pi-": 2}
+    )
+    assert dc_flatten.bf == approx(0.5 * (0.0124 ** 2) * (0.692 ** 2))
+
+
+def test_DecayChain_string_repr(dc):
     assert str(dc) == "<DecayChain: D0 -> K_S0 pi0 (2 sub-decays), BF=0.0124>"